[fix] youtube api からのデータの取得・加工をmodelで行うように修正
データとデータに関わるロジックをモデルに統一するため
